requirements.txt
altgraph==0.17
appdirs==1.4.3
#   apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
auto-py-to-exe==2.7.8
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
bottle==0.12.18
bottle-websocket==0.2.9
#Brlapi==0.7.0
Brotli==1.0.9
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
#command-not-found==0.3
cryptography==2.8
#cupshelpers==1.0
#dbus-python==1.2.16
decorator==4.4.2
defer==1.0.4
defusedxml==0.6.0
distlib==0.3.0
distro==1.5.0
#distro-info===0.10ubuntu1
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-twitter-feed==0.2
djangorestframework==3.12.1
docker==4.3.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
#duplicity==0.8.17.dev8
Eel==0.12.4
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flake8==3.8.3
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
httplib2==0.14.0
idna==2.10
import==0.0.1
import-ipynb==0.1.3
importlib-metadata==1.5.0
ipykernel==5.3.4
ipynb==0.5.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.2
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==18.0.1
#language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
# louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
nltk==3.2.1
notebook==6.1.1
numpy==1.19.2
oauth==1.0.1
oauth-python-twitter==1.0
oauth2client==4.1.2
oauthlib==2.1.0
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.1
pbr==5.5.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pip-init==0.3.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psycopg2==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.8
#pycups==1.9.73
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.3.1
#PyGObject==3.36.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.16.0
PySocks==1.7.1
#python-apt==0.7.8
python-dateutil==2.7.3
python-debian===0.1.38
python-dotenv==0.14.0
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-oauth2==1.1.1
python-social-auth==0.2.1
python-twitter==3.5
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.9.27
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.1.0
requests-unixsocket==0.2.0
rsa==4.0
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SQLAlchemy==1.3.12
sqlparse==0.3.1
stevedore==3.2.2
stream-django==1.6.2
stream-python==4.0.0
#systemd-python==234
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.50.0
traitlets==4.3.3
tweepy==3.9.0
#ubuntu-advantage-tools==20.3
#ubuntu-drivers-common==0.0.0
#ufw==0.36
#ujson==3.2.0
#unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.10
#usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.53.0
whichcraft==0.6.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
#xkit==0.0.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.1.0
psycopg2